move packet into a different file
